+++ R/chat.R
@@ -22,11 +22,7 @@ Chat <- R6::R6Class(
   "Chat",
   public = list(
     #' @param provider A provider object.
-    #' @param turns An unnamed list of turns to start the chat with (i.e.,
-    #'   continuing a previous conversation). If `NULL` or zero-length list, the
-    #'   conversation begins from scratch.
-    #' @param seed Optional integer seed that ChatGPT uses to try and make output
-    #'   more reproducible.
+    #' @param system_prompt System prompt to start the conversation with.
     #' @param echo One of the following options:
     #'   * `none`: don't emit any output (default when running in a function).
     #'   * `text`: echo text output as it streams in (default when running at
@@ -34,10 +30,10 @@ Chat <- R6::R6Class(
     #'   * `all`: echo all input and output.
     #'
     #'  Note this only affects the `chat()` method.
-    initialize = function(provider, turns, seed = NULL, echo = "none") {
+    initialize = function(provider, system_prompt = NULL, echo = "none") {
       private$provider <- provider
-      private$.turns <- turns %||% list()
       private$echo <- echo
+      self$set_system_prompt(system_prompt)
     },
 
     #' @description Retrieve the turns that have been sent and received so far
@@ -94,9 +90,13 @@ Chat <- R6::R6Class(
     },
 
     #' @description Update the system prompt
-    #' @param value A string giving the new system prompt
+    #' @param value A character vector giving the new system prompt
     set_system_prompt = function(value) {
-      check_string(value, allow_null = TRUE)
+      check_character(value, allow_null = TRUE)
+      if (length(value) > 1) {
+        value <- paste(value, collapse = "\n\n")
+      }
+
       # Remove prompt, if present
       if (private$has_system_prompt()) {
         private$.turns <- private$.turns[-1]
@@ -455,7 +455,7 @@ Chat <- R6::R6Class(
   private = list(
     provider = NULL,
 
-    .turns = NULL,
+    .turns = list(),
     echo = NULL,
     tools = list(),