+++ R/provider-openai.R
@@ -14,9 +14,6 @@ NULL
 #' [developer platform](https://platform.openai.com).
 #'
 #' @param system_prompt A system prompt to set the behavior of the assistant.
-#' @param turns A list of [Turn]s to start the chat with (i.e., continuing a
-#'   previous conversation). If not provided, the conversation begins from
-#'   scratch.
 #' @param base_url The base URL to the endpoint; the default uses OpenAI.
 #' @param api_key `r api_key_param("OPENAI_API_KEY")`
 #' @param model The model to use for the chat. The default, `NULL`, will pick
@@ -48,7 +45,6 @@ NULL
 #' chat$chat("Tell me three funny jokes about statisticians")
 chat_openai <- function(
   system_prompt = NULL,
-  turns = NULL,
   base_url = "https://api.openai.com/v1",
   api_key = openai_key(),
   model = NULL,
@@ -57,7 +53,6 @@ chat_openai <- function(
   api_args = list(),
   echo = c("none", "output", "all")
 ) {
-  turns <- normalize_turns(turns, system_prompt)
   model <- set_default(model, "gpt-4o")
   echo <- check_echo(echo)
 
@@ -79,7 +74,7 @@ chat_openai <- function(
     extra_args = api_args,
     api_key = api_key
   )
-  Chat$new(provider = provider, turns = turns, echo = echo)
+  Chat$new(provider = provider, system_prompt = system_prompt, echo = echo)
 }
 
 chat_openai_test <- function(..., model = "gpt-4o-mini", params = NULL) {